Bring delete role endpoint in line with restful standards
